+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.2.0
 
 - **_Breaking Change_**: rename `LogProxy.disable_lambda_telemetry_log_fd` to `LogProxy.disable_lambda_telemetry_log_fd_for_handler`.
+- **_Breaking Change_**: rename `Processor` and `ProcessorBuilder` into `SimpleProcessor` and `SimpleProcessorBuilder`.
+- **_Breaking Change_**: add trait `Processor`, make `LogProxy` generic.
 - **_Breaking Change_**: `Sink::write_line` and `Processor::process` need a timestamp as the second argument.
 - Feat: add `LogProxy::buffer_size`, store lines in a buffer and record the in-buffer timestamp.
 - ~~Feat: add `LogProxy::suppression_timeout_ms` to customize the suppression timeout.~~